Always regenerate the cross app import graph (#18533)
* update cross-app import job

* update regenerate cross-app import script

* The cross-app import graph was automatically updated. Don't delete!

* update ignored files for testing

* test - update md file

* ignore another file for testing purposes

* test - change an app file

* clean up

* cancel worflow  after updated graph is pushed to branch

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* step must use stefanzweifel/git-auto-commit-action@v4

* test - force change in graph

* wip - cancel workflow if graph was updated

* The cross-app import graph was automatically updated. Don't delete!

* wip - cancel workflow if graph was updated

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* wip - cancel workflow if graph was updated

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* wip - cancel workflow if graph was updated

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* echo changes_detected

* don't echo changes_detected

* checkout repo with vsp bot token

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* clean up

* test - force change in graph

* The cross-app import graph was automatically updated. Don't delete!

* update step name

* cross-app import clean up

* The cross-app import graph was automatically updated. Don't delete!

* remove regenerate-cross-app-import-graph job

* delete config/cross_app_import_graph.json

* move graph code into select-cypress-tests.js

* delete create-cross-app-import-graph.js

* fully delete regenerate-cross-app-import-graph job :-)

* comment-out concurency block

* ignore files during testing

* log the graph during testing

* change ask-a-question file

* test - change caregivers file

* clean up
