+++ .github/workflows/continuous-integration.yml
@@ -430,70 +430,9 @@ jobs:
       - name: Annotate Stylelint results
         if: ${{ always() }}
         run: node ./script/github-actions/stylelint-annotation-format.js
-  
-  check-app-dependency-graph:
-    name: Check App Dependency Graph
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        if: ${{ github.ref != 'refs/heads/master' }}
-        uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-
-      - name: Get Node version
-        if: ${{ github.ref != 'refs/heads/master' }}
-        id: get-node-version
-        run: echo ::set-output name=NODE_VERSION::$(cat .nvmrc)
-
-      - name: Setup Node
-        if: ${{ github.ref != 'refs/heads/master' }}
-        uses: actions/setup-node@v2
-        with:
-          node-version: ${{ steps.get-node-version.outputs.NODE_VERSION }}
-
-      - name: Cache dependencies
-        if: ${{ github.ref != 'refs/heads/master' }}
-        id: cache-dependencies
-        uses: actions/cache@v2
-        with:
-          path: |
-            .cache/yarn
-            node_modules
-          key: ${{ steps.get-node-version.outputs.NODE_VERSION }}-${{ hashFiles('yarn.lock') }}
-          restore-keys: ${{ steps.get-node-version.outputs.NODE_VERSION }}-
-
-      - name: Install dependencies
-        if: ${{ github.ref != 'refs/heads/master' }}
-        uses: nick-invision/retry@v2
-        with:
-          command: yarn install --frozen-lockfile --prefer-offline --production=false
-          max_attempts: 3
-          timeout_minutes: 7
-        env:
-          YARN_CACHE_FOLDER: .cache/yarn
-
-      - name: Run git diff and set DIFF_RESULTS
-        if: ${{ github.ref != 'refs/heads/master' }}
-        id: git-diff
-        run: echo ::set-output name=DIFF_RESULTS::$(git diff ${{ github.event.before }} ${{ github.event.after }} --line-prefix=new_line --unified=0)
-
-      - name: Set IS_GRAPH_UPDATED and update the cross-app import graph if necessary
-        if: ${{ github.ref != 'refs/heads/master' }}
-        run: node script/github-actions/create-cross-app-import-graph.js
-        env:
-          DIFF_RESULTS: ${{ steps.git-diff.outputs.DIFF_RESULTS }}
-
-      - name: Push the updated cross-app import graph
-        if: ${{ github.ref != 'refs/heads/master' && env.IS_GRAPH_UPDATED == 'true' }}
-        uses: stefanzweifel/git-auto-commit-action@v4
-        with:
-          commit_message: update graph
 
   cypress-tests-prep:
     name: Prep for Cypress Tests
-    needs: check-app-dependency-graph
     runs-on: ubuntu-latest
     outputs:
       num_containers: ${{ steps.containers.outputs.num }}
